requirements.txt
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
arxiv==1.4.8
google-search-results==2.4.2
pytz==2023.3
